@@ -26,7 +26,8 @@ from __future__ import annotations
 
 import dataclasses
 import re
-from typing import TYPE_CHECKING, Any, Dict, Generator, List, Optional, Tuple, Union
+from collections.abc import Generator
+from typing import TYPE_CHECKING, Any, Optional, Union
 
 from .abc import ReconstructAble
 from .http import HTTPClientT
@@ -35,7 +36,7 @@ from .utils import parse_time, simple_repr
 if TYPE_CHECKING:
     import datetime
 
-__all__: Tuple[str, ...] = ("Aes", "DynamicKey", "Version")
+__all__: tuple[str, ...] = ("Aes", "DynamicKey", "Version")
 
 VERSION_REGEX: re.Pattern[str] = re.compile(r"(?P<version>[0-9]{2})\.(?P<subversion>[0-9]{2})")
 
@@ -88,7 +89,7 @@ class Version:
 
 
 @simple_repr
-class Aes(ReconstructAble[Dict[str, Any], HTTPClientT]):
+class Aes(ReconstructAble[dict[str, Any], HTTPClientT]):
     """
     .. attributetable:: fortnite_api.Aes
 
@@ -143,7 +144,7 @@ class Aes(ReconstructAble[Dict[str, Any], HTTPClientT]):
         All current dynamic keys
     """
 
-    __slots__: Tuple[str, ...] = (
+    __slots__: tuple[str, ...] = (
         "main_key",
         "build",
         "version",
@@ -151,7 +152,7 @@ class Aes(ReconstructAble[Dict[str, Any], HTTPClientT]):
         "dynamic_keys",
     )
 
-    def __init__(self, *, data: Dict[str, Any], http: HTTPClientT):
+    def __init__(self, *, data: dict[str, Any], http: HTTPClientT):
         super().__init__(data=data, http=http)
 
         self.main_key: Optional[str] = data.get("mainKey")
@@ -164,7 +165,7 @@ class Aes(ReconstructAble[Dict[str, Any], HTTPClientT]):
             major, minor = version_info[0]
             self.version = Version(major=int(major), minor=int(minor))
 
-        self.dynamic_keys: List[DynamicKey[HTTPClientT]] = [
+        self.dynamic_keys: list[DynamicKey[HTTPClientT]] = [
             DynamicKey(data=entry, http=http) for entry in data.get("dynamicKeys", [])
         ]
         self.updated: datetime.datetime = parse_time(data["updated"])
@@ -189,7 +190,7 @@ class Aes(ReconstructAble[Dict[str, Any], HTTPClientT]):
 
 
 @simple_repr
-class DynamicKey(ReconstructAble[Dict[str, Any], HTTPClientT]):
+class DynamicKey(ReconstructAble[dict[str, Any], HTTPClientT]):
     """
     .. attributetable:: fortnite_api.DynamicKey
 
@@ -227,9 +228,9 @@ class DynamicKey(ReconstructAble[Dict[str, Any], HTTPClientT]):
         The key.
     """
 
-    __slots__: Tuple[str, ...] = ("pak_filename", "pak_guid", "key")
+    __slots__: tuple[str, ...] = ("pak_filename", "pak_guid", "key")
 
-    def __init__(self, *, data: Dict[str, Any], http: HTTPClientT):
+    def __init__(self, *, data: dict[str, Any], http: HTTPClientT):
         super().__init__(data=data, http=http)
 
         self.pak_filename: str = data["pakFilename"]
