@@ -22,19 +22,29 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-from typing import Any, Dict, Tuple
+from collections.abc import Generator
+from typing import Any
 
 from .abc import ReconstructAble
-from .cosmetics import CosmeticBr, CosmeticCar, CosmeticInstrument, CosmeticLegoKit, CosmeticTrack, VariantBean, VariantLego
+from .cosmetics import (
+    Cosmetic,
+    CosmeticBr,
+    CosmeticCar,
+    CosmeticInstrument,
+    CosmeticLegoKit,
+    CosmeticTrack,
+    VariantBean,
+    VariantLego,
+)
 from .http import HTTPClientT
 from .proxies import TransformerListProxy
 from .utils import get_with_fallback, simple_repr
 
-__all__: Tuple[str, ...] = ("CosmeticsAll",)
+__all__: tuple[str, ...] = ("CosmeticsAll",)
 
 
 @simple_repr
-class CosmeticsAll(ReconstructAble[Dict[str, Any], HTTPClientT]):
+class CosmeticsAll(ReconstructAble[dict[str, Any], HTTPClientT]):
     """
     .. attributetable:: fortnite_api.CosmeticsAll
 
@@ -86,7 +96,7 @@ class CosmeticsAll(ReconstructAble[Dict[str, Any], HTTPClientT]):
         A list of all lego kit cosmetics.
     """
 
-    __slots__: Tuple[str, ...] = (
+    __slots__: tuple[str, ...] = (
         "br",
         "tracks",
         "instruments",
@@ -96,7 +106,7 @@ class CosmeticsAll(ReconstructAble[Dict[str, Any], HTTPClientT]):
         "beans",
     )
 
-    def __init__(self, *, data: Dict[str, Any], http: HTTPClientT) -> None:
+    def __init__(self, *, data: dict[str, Any], http: HTTPClientT) -> None:
         super().__init__(data=data, http=http)
 
         _br = get_with_fallback(data, "br", list)
@@ -141,29 +151,22 @@ class CosmeticsAll(ReconstructAble[Dict[str, Any], HTTPClientT]):
             lambda x: VariantBean(data=x, http=self._http),
         )
 
-    def __iter__(self):
-        for br in self.br:
-            yield br
+    def __iter__(self) -> Generator[Cosmetic[dict[str, Any], HTTPClientT], None, None]:
+        yield from self.br
 
-        for track in self.tracks:
-            yield track
+        yield from self.tracks
 
-        for instrument in self.instruments:
-            yield instrument
+        yield from self.instruments
 
-        for car in self.cars:
-            yield car
+        yield from self.cars
 
-        for lego in self.lego:
-            yield lego
+        yield from self.lego
 
-        for bean in self.beans:
-            yield bean
+        yield from self.beans
 
-        for lego_kit in self.lego_kits:
-            yield lego_kit
+        yield from self.lego_kits
 
-    def __len__(self):
+    def __len__(self) -> int:
         return (
             len(self.br)
             + len(self.tracks)
