@@ -24,7 +24,8 @@ SOFTWARE.
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Coroutine, Generic, Optional, Tuple, Union, overload
+from collections.abc import Coroutine
+from typing import TYPE_CHECKING, Any, Generic, Optional, Union, overload
 
 from typing_extensions import Self
 
@@ -35,7 +36,7 @@ if TYPE_CHECKING:
     from .http import HTTPClient, SyncHTTPClient
 
 
-__all__: Tuple[str, ...] = ('Asset',)
+__all__: tuple[str, ...] = ('Asset',)
 
 
 class _AssetRoute(Route):
@@ -63,7 +64,7 @@ class Asset(Generic[HTTPClientT]):
             icon: bytes = await images.icon.read()
     """
 
-    __slots__: Tuple[str, ...] = ('_http', '_url', '_max_size', '_size')
+    __slots__: tuple[str, ...] = ('_http', '_url', '_max_size', '_size')
 
     def __init__(self, *, http: HTTPClientT, url: str, max_size: Optional[int] = MISSING, size: int = MISSING) -> None:
         self._http: HTTPClientT = http
@@ -89,7 +90,7 @@ class Asset(Generic[HTTPClientT]):
         return hash(self.url)
 
     def __repr__(self) -> str:
-        return '<Asset url={0.url!r}>'.format(self)
+        return f'<Asset url={self.url!r}>'
 
     @property
     def url(self) -> str:
