@@ -27,11 +27,12 @@ from __future__ import annotations
 import datetime
 import functools
 import inspect
-from typing import Any, Callable, List, Literal, Optional, Tuple, TypeVar, Union, cast, overload
+from collections.abc import Coroutine
+from typing import Any, Callable, Literal, Optional, TypeVar, Union, cast, overload
 
 import aiohttp
 import requests
-from typing_extensions import Concatenate, Coroutine, ParamSpec, Self
+from typing_extensions import Concatenate, ParamSpec, Self
 
 from .aes import Aes
 from .all import CosmeticsAll
@@ -52,7 +53,7 @@ from .shop import Shop
 from .stats import BrPlayerStats
 from .utils import MISSING, _transform_dict_for_get_request, copy_doc, remove_prefix
 
-__all__: Tuple[str, ...] = (
+__all__: tuple[str, ...] = (
     'Client',
     'SyncClient',
 )
@@ -241,7 +242,7 @@ class Client:
 
     async def fetch_cosmetics_br(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[CosmeticBr]:
+    ) -> list[CosmeticBr]:
         """|coro|
 
         Fetches all Battle Royale cosmetics available in Fortnite.
@@ -275,7 +276,7 @@ class Client:
 
     async def fetch_cosmetics_cars(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[CosmeticCar]:
+    ) -> list[CosmeticCar]:
         """|coro|
 
         Fetches all Car cosmetics available in Fortnite.
@@ -309,7 +310,7 @@ class Client:
 
     async def fetch_cosmetics_instruments(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[CosmeticInstrument]:
+    ) -> list[CosmeticInstrument]:
         """|coro|
 
         Fetches all instrument cosmetics available in Fortnite.
@@ -343,7 +344,7 @@ class Client:
 
     async def fetch_cosmetics_lego_kits(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[CosmeticLegoKit]:
+    ) -> list[CosmeticLegoKit]:
         """|coro|
 
         Fetches all lego kit cosmetics available in Fortnite.
@@ -377,7 +378,7 @@ class Client:
 
     async def fetch_variants_lego(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[VariantLego]:
+    ) -> list[VariantLego]:
         """|coro|
 
         Fetches all lego cosmetic variants available in Fortnite.
@@ -412,7 +413,7 @@ class Client:
 
     async def fetch_variants_beans(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[VariantBean]:
+    ) -> list[VariantBean]:
         """|coro|
 
         Fetches all bean cosmetic variants available in Fortnite. For more information
@@ -448,7 +449,7 @@ class Client:
 
     async def fetch_cosmetics_tracks(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[CosmeticTrack]:
+    ) -> list[CosmeticTrack]:
         """|coro|
 
         Fetches all audio track cosmetics available in Fortnite.
@@ -597,7 +598,7 @@ class Client:
         added_since: Optional[datetime.datetime] = ...,
         unseen_for: Optional[int] = ...,
         last_appearance: Optional[datetime.datetime] = ...,
-    ) -> List[CosmeticBr]: ...
+    ) -> list[CosmeticBr]: ...
 
     @overload
     async def search_br_cosmetics(
@@ -640,7 +641,7 @@ class Client:
         last_appearance: Optional[datetime.datetime] = ...,
     ) -> CosmeticBr: ...
 
-    async def search_br_cosmetics(self, **kwargs: Any) -> Union[CosmeticBr, List[CosmeticBr]]:
+    async def search_br_cosmetics(self, **kwargs: Any) -> Union[CosmeticBr, list[CosmeticBr]]:
         """|coro|
 
         Searches all Battle Royale cosmetics available in Fortnite and returns
@@ -774,7 +775,7 @@ class Client:
         return Aes(data=data, http=self.http)
 
     # BANNERS
-    async def fetch_banners(self, *, language: Optional[GameLanguage] = MISSING) -> List[Banner]:
+    async def fetch_banners(self, *, language: Optional[GameLanguage] = MISSING) -> list[Banner]:
         """|coro|
 
         Fetch all banners available in Fortnite.
@@ -798,7 +799,7 @@ class Client:
             lambda x: Banner(data=x, http=self.http),
         )
 
-    async def fetch_banner_colors(self) -> List[BannerColor]:
+    async def fetch_banner_colors(self) -> list[BannerColor]:
         """|coro|
 
         Fetch all banner colors available in Fortnite.
@@ -939,7 +940,7 @@ class Client:
 
     # PLAYLISTS
 
-    async def fetch_playlists(self, /, *, language: Optional[GameLanguage] = MISSING) -> List[Playlist]:
+    async def fetch_playlists(self, /, *, language: Optional[GameLanguage] = MISSING) -> list[Playlist]:
         """|coro|
 
         Fetches a list of current playlists available in Fortnite.
@@ -1103,7 +1104,7 @@ class Client:
     # BETA METHODS
 
     @beta_method
-    async def beta_fetch_new_display_assets(self) -> List[NewDisplayAsset]:
+    async def beta_fetch_new_display_assets(self) -> list[NewDisplayAsset]:
         """|coro|
 
         Fetches all the new display assets available in Fortnite.
@@ -1153,7 +1154,7 @@ class Client:
         )
 
     @beta_method
-    async def beta_fetch_material_instances(self) -> List[MaterialInstance]:
+    async def beta_fetch_material_instances(self) -> list[MaterialInstance]:
         """|coro|
 
         Fetches all the material instances available in Fortnite.
@@ -1300,7 +1301,7 @@ class SyncClient:
     @copy_doc(Client.fetch_cosmetics_br)
     def fetch_cosmetics_br(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[CosmeticBr[SyncHTTPClient]]:
+    ) -> list[CosmeticBr[SyncHTTPClient]]:
         data = self.http.get_cosmetics_br(
             language=self._resolve_default_language_value(language),
             response_flags=self._resolve_response_flags_value(response_flags),
@@ -1313,7 +1314,7 @@ class SyncClient:
     @copy_doc(Client.fetch_cosmetics_cars)
     def fetch_cosmetics_cars(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[CosmeticCar[SyncHTTPClient]]:
+    ) -> list[CosmeticCar[SyncHTTPClient]]:
         data = self.http.get_cosmetics_cars(
             language=self._resolve_default_language_value(language),
             response_flags=self._resolve_response_flags_value(response_flags),
@@ -1326,7 +1327,7 @@ class SyncClient:
     @copy_doc(Client.fetch_cosmetics_instruments)
     def fetch_cosmetics_instruments(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[CosmeticInstrument[SyncHTTPClient]]:
+    ) -> list[CosmeticInstrument[SyncHTTPClient]]:
         data = self.http.get_cosmetics_instruments(
             language=self._resolve_default_language_value(language),
             response_flags=self._resolve_response_flags_value(response_flags),
@@ -1339,7 +1340,7 @@ class SyncClient:
     @copy_doc(Client.fetch_cosmetics_lego_kits)
     def fetch_cosmetics_lego_kits(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[CosmeticLegoKit[SyncHTTPClient]]:
+    ) -> list[CosmeticLegoKit[SyncHTTPClient]]:
         data = self.http.get_cosmetics_lego_kits(
             language=self._resolve_default_language_value(language),
             response_flags=self._resolve_response_flags_value(response_flags),
@@ -1352,7 +1353,7 @@ class SyncClient:
     @copy_doc(Client.fetch_variants_lego)
     def fetch_variants_lego(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[VariantLego[SyncHTTPClient]]:
+    ) -> list[VariantLego[SyncHTTPClient]]:
         data = self.http.get_cosmetics_lego(
             language=self._resolve_default_language_value(language),
             response_flags=self._resolve_response_flags_value(response_flags),
@@ -1365,7 +1366,7 @@ class SyncClient:
     @copy_doc(Client.fetch_variants_beans)
     def fetch_variants_beans(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[VariantBean[SyncHTTPClient]]:
+    ) -> list[VariantBean[SyncHTTPClient]]:
         data = self.http.get_cosmetics_beans(
             language=self._resolve_default_language_value(language),
             response_flags=self._resolve_response_flags_value(response_flags),
@@ -1378,7 +1379,7 @@ class SyncClient:
     @copy_doc(Client.fetch_cosmetics_tracks)
     def fetch_cosmetics_tracks(
         self, *, language: Optional[GameLanguage] = MISSING, response_flags: Optional[ResponseFlags] = MISSING
-    ) -> List[CosmeticTrack[SyncHTTPClient]]:
+    ) -> list[CosmeticTrack[SyncHTTPClient]]:
         data = self.http.get_cosmetics_tracks(
             language=self._resolve_default_language_value(language),
             response_flags=self._resolve_response_flags_value(response_flags),
@@ -1455,7 +1456,7 @@ class SyncClient:
         added_since: Optional[datetime.datetime] = ...,
         unseen_for: Optional[int] = ...,
         last_appearance: Optional[datetime.datetime] = ...,
-    ) -> List[CosmeticBr[SyncHTTPClient]]: ...
+    ) -> list[CosmeticBr[SyncHTTPClient]]: ...
 
     @overload
     def search_br_cosmetics(
@@ -1499,7 +1500,7 @@ class SyncClient:
     ) -> CosmeticBr[SyncHTTPClient]: ...
 
     @copy_doc(Client.search_br_cosmetics)
-    def search_br_cosmetics(self, **kwargs: Any) -> Union[CosmeticBr[SyncHTTPClient], List[CosmeticBr[SyncHTTPClient]]]:
+    def search_br_cosmetics(self, **kwargs: Any) -> Union[CosmeticBr[SyncHTTPClient], list[CosmeticBr[SyncHTTPClient]]]:
         multiple = kwargs.pop('multiple')
 
         kwargs['language'] = self._resolve_default_language_value(kwargs.pop('language', MISSING))
@@ -1530,7 +1531,7 @@ class SyncClient:
 
     # BANNERS
     @copy_doc(Client.fetch_banners)
-    def fetch_banners(self, *, language: Optional[GameLanguage] = MISSING) -> List[Banner[SyncHTTPClient]]:
+    def fetch_banners(self, *, language: Optional[GameLanguage] = MISSING) -> list[Banner[SyncHTTPClient]]:
         data = self.http.get_banners(language=self._resolve_default_language_value(language))
         return TransformerListProxy(
             data,
@@ -1538,7 +1539,7 @@ class SyncClient:
         )
 
     @copy_doc(Client.fetch_banner_colors)
-    def fetch_banner_colors(self) -> List[BannerColor[SyncHTTPClient]]:
+    def fetch_banner_colors(self) -> list[BannerColor[SyncHTTPClient]]:
         data = self.http.get_banner_colors()
         return TransformerListProxy(
             data,
@@ -1579,7 +1580,7 @@ class SyncClient:
     # PLAYLISTS
 
     @copy_doc(Client.fetch_playlists)
-    def fetch_playlists(self, /, *, language: Optional[GameLanguage] = MISSING) -> List[Playlist[SyncHTTPClient]]:
+    def fetch_playlists(self, /, *, language: Optional[GameLanguage] = MISSING) -> list[Playlist[SyncHTTPClient]]:
         data = self.http.get_playlists(language=self._resolve_default_language_value(language))
         return TransformerListProxy(
             data,
@@ -1625,7 +1626,7 @@ class SyncClient:
 
     @copy_doc(Client.beta_fetch_new_display_assets)
     @beta_method
-    def beta_fetch_new_display_assets(self) -> List[NewDisplayAsset[SyncHTTPClient]]:
+    def beta_fetch_new_display_assets(self) -> list[NewDisplayAsset[SyncHTTPClient]]:
         data = self.http.beta_get_new_display_assets()
 
         return TransformerListProxy(
@@ -1635,7 +1636,7 @@ class SyncClient:
 
     @copy_doc(Client.beta_fetch_material_instances)
     @beta_method
-    def beta_fetch_material_instances(self) -> List[MaterialInstance[SyncHTTPClient]]:
+    def beta_fetch_material_instances(self) -> list[MaterialInstance[SyncHTTPClient]]:
         data = self.http.beta_get_material_instances()
 
         return TransformerListProxy(
