@@ -25,11 +25,10 @@ SOFTWARE.
 from __future__ import annotations
 
 import enum
-from typing import Tuple, Type
 
 from typing_extensions import Self
 
-__all__: Tuple[str, ...] = (
+__all__: tuple[str, ...] = (
     'KeyFormat',
     'GameLanguage',
     'MatchMethod',
@@ -369,7 +368,7 @@ class CosmeticCompatibleMode(enum.Enum):
     ALL = 'max'
 
     @classmethod
-    def _from_str(cls: Type[Self], string: str) -> Self:
+    def _from_str(cls: type[Self], string: str) -> Self:
         # The Epic Games API uses both "CosmeticCompatibleMode" and "CosmeticCompatibleModeLegacy" enums
         # with the same values, so we need to handle both.
         # To easily handle this, we'll remove the "ECosmeticCompatibleMode::" or "ECosmeticCompatibleModeLegacy::" prefix.
@@ -434,7 +433,7 @@ class ProductTag(enum.Enum):
     ALL = 'max'
 
     @classmethod
-    def _from_str(cls: Type[Self], string: str) -> Self:
+    def _from_str(cls: type[Self], string: str) -> Self:
         # The Epic Games API "Product" enums contains both lower case and capitalized values, so we need to handle both.
         # To easily handle this, we'll remove the "Product." prefix and convert it to lowercase.
         trimmed = string.split('.')[-1]
